% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_estimator.R
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimate regression coefficients beta under various methods}
\usage{
estimate_beta(
  y,
  x,
  method = c("OLS", "Huber", "Profile", "Adj profile", "IJ", "Marginalized IJ",
    "Marginalized Fisher"),
  omega_init = 0.5,
  beta_init = NULL,
  sigma_init = NULL,
  control = list(maxit = 10000)
)
}
\arguments{
\item{y}{Response vector}

\item{x}{Design matrix (including intercept)}

\item{method}{Character or numeric; one of:
"OLS", "Huber", "profile", "adj_profile", "IJ", "Marginalized IJ", "Marginalized Fisher",
or a non-negative integer for fixed nu approach.}

\item{omega_init}{Initial value for omega (default = 0.5)}

\item{beta_init}{Optional initial beta values}

\item{sigma_init}{Optional initial sigma}

\item{control}{List of control parameters for optim}
}
\value{
A list with components:
\describe{
\item{method}{The method used}
\item{nu_hat}{Estimated or fixed degrees of freedom (NA for OLS/Huber)}
\item{omega_hat}{Transformed 1/nu parameter (NA for OLS/Huber)}
\item{beta_hat}{Estimated coefficients}
\item{sigma_hat}{Estimated sigma (residual SD)}
\item{success_nu}{Convergence code for nu estimation. A value of
0 indicates successful convergence. (NA for fixed nu/OLS/Huber)}
\item{success_beta}{Convergence code for beta optimization. A value of
0 indicates successful convergence. (NA for OLS/Huber)}
}
}
\description{
This function estimates regression coefficients (\eqn{\beta}) and scale (\eqn{\sigma})
using one of the following approaches:
\itemize{
\item Ordinary least squares ("OLS")
\item Huber regression ("Huber")
\item Fixed degrees of freedom t-regression (method = non-negative integer)
\item Estimated degrees of freedom using one of five methods:
"Profile", "Adj profile", "IJ", "Marginalized IJ", or "Marginalized Fisher"
}
}
