% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_estimator.R
\name{estimate_nu_adj_profile}
\alias{estimate_nu_adj_profile}
\title{Adjusted-profile estimator (uses adjustment term based on observed j matrix)}
\usage{
estimate_nu_adj_profile(y, x, omega_init = 1/2)
}
\arguments{
\item{y}{Numeric response vector.}

\item{x}{Numeric design matrix (including an intercept column).}

\item{omega_init}{Starting value for \code{omega} optimization.
\eqn{\omega = 1 / \nu}.If simulated t-error data is used,
use true 1/\eqn{\nu} for quicker convergence.}
}
\value{
list with optimized \eqn{\omega}, optimized \eqn{\nu} and convergence status
}
\description{
Adjusted-profile estimator (uses adjustment term based on observed j matrix)
}
