% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_estimator.R
\name{estimate_nu_IJ}
\alias{estimate_nu_IJ}
\title{Independent Jeffrey's estimator}
\usage{
estimate_nu_IJ(y, x, omega_init = 1/2)
}
\arguments{
\item{y}{Numeric vector. The response variable.}

\item{x}{Numeric matrix. The design matrix (should include an intercept column).}

\item{omega_init}{Numeric. Initial value for \eqn{\omega = 1 / \nu}.
If simulated t-error data is used, setting this to the true value of \eqn{1/\nu}
can improve convergence speed.}
}
\value{
A list with the following components:
\describe{
\item{omega}{The optimized value of \eqn{\omega}.}
\item{nu}{The corresponding value of \eqn{\nu = 1 / \omega}.}
\item{convergence}{An integer convergence code from \code{\link[stats]{optim}}.
A value of 0 indicates successful convergence.}
}
}
\description{
Computes the Independent Jeffrey's approach estimator for the degrees of freedom
parameter \eqn{\nu} via optimization over \eqn{\omega = 1 / \nu}.
}
