% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_estimator.R
\name{optimize_omega}
\alias{optimize_omega}
\title{Optimize \eqn{\omega} parameters for a given target function}
\usage{
optimize_omega(omega_init, target_fn)
}
\arguments{
\item{omega_init}{Numeric vector of initial values for the \eqn{\omega} parameters.}

\item{target_fn}{A function to be minimized, taking \code{omega} as input
and returning a scalar objective value.}
}
\value{
A list containing:
\describe{
\item{omega}{The optimized \eqn{\omega} parameter vector.}
\item{convergence}{Convergence code returned by \code{optim}. A value of
0 indicates successful convergence.}
\item{value}{The final value of the objective function at the optimum.}
}
}
\description{
This function performs numerical optimization over the
omega parameter vector using the BFGS algorithm. It is primarily designed
for minimizing an omega-based objective function within the package's internal
estimation routines.
}
\keyword{internal}
